+++ config/dummyLikes.ts
