+++ views/profil/BookmarkSection/Bookmark.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image"
 import Link from "next/link"
-import dummyBookmarks from "@constants/dummyBookmarks"
+import dummyBookmarks from "../../../config/dummyBookmarks"
 
 const BookmarkSection = () => {
   return (
